This commit marks a change in how we generate code from our CDM XSD files.  In the past we generated one hxx/cxx file for each type in the XSD files but to speed up compilation we have switched to one file per XSD file.
This required us to modify all of our files that referenced hxx/cxx
files to reference the new correct file. While I was at it I sorted most
includes and placed the corisponding header for each cpp as the first
include where I could.

Addiitonal changes include changeing most empty constructors in SETypes
to = default and adding override markers where I noticed the. We also no
longer use stdafx.h while compiling and so many headers make direct
reference to COmmonDataModel.h and Biogears.h which were previously
bundeled in these precompiled headers.

Finally the UCCommon header is no longer referenced as it had a lot of
circular includes and many CPP files did not einclude there own header
manually.

This work was squash commited to reduce the noice as a lot of testing
went in to making sure the final changes were supported on all platforms
as compiler differences were not aparrent while making the header
changes.

+++ projects/biogears/include/biogears/cdm/circuit/SECircuitCalculator.inl
@@ -15,7 +15,6 @@ specific language governing permissions and limitations under the License.
 #include <biogears/cdm/circuit/SECircuitNode.h>
 #include <biogears/cdm/circuit/SECircuitPath.h>
 #include <biogears/cdm/properties/SEScalar.h>
-#include <biogears/cdm/stdafx.h>
 #include <biogears/cdm/utils/GeneralMath.h>
 
 #include <Eigen/Dense>
@@ -149,7 +148,7 @@ void SECircuitCalculator<CIRCUIT_CALCULATOR_TYPES>::ParseIn()
   }
 
   size_t idx = 0;
-  int numNodes = m_circuit->GetNodes().size() - m_circuit->GetReferenceNodes().size();
+  size_t numNodes = m_circuit->GetNodes().size() - m_circuit->GetReferenceNodes().size();
   m_potentialSources.clear();
   for (PathType* p : m_circuit->GetPaths()) {
     //Set aside the pressure sources, since the Flow through them will be directly solved by adding them to the bottom of the matrix.
@@ -159,8 +158,8 @@ void SECircuitCalculator<CIRCUIT_CALCULATOR_TYPES>::ParseIn()
     }
   }
 
-  int numSources = m_potentialSources.size();
-  int numVars = numNodes + numSources;
+  size_t numSources = m_potentialSources.size();
+  size_t numVars = numNodes + numSources;
   //Set the size of the matrix and initialize all elements to zero - we'll populate it later.
   //subtract the known reference - we don't need to solve for that pressure
   //(and it has to be known or we'll have too many unknowns).
