+++ projects/biogears/src/cdm/properties/SEHistogram.cpp
@@ -9,14 +9,9 @@ the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
 CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 **************************************************************************************/
-
 #include <biogears/cdm/properties/SEHistogram.h>
+
 #include <biogears/cdm/properties/SEScalar.h> //Utils
-#include <biogears/cdm/stdafx.h>
-#include <biogears/schema/DoubleArray.hxx>
-#include <biogears/schema/DoubleList.hxx>
-#include <biogears/schema/HistogramData.hxx>
-#include <biogears/schema/Properties.hxx>
 
 static std::stringstream err;
 
@@ -88,14 +83,14 @@ void SEHistogram::Unload(CDM::HistogramData& data) const
 unsigned int SEHistogram::NumberOfBins() const
 {
   if (IsValid())
-    return m_Dependent.size();
+    return static_cast<unsigned int>(m_Dependent.size());
   return 0;
 }
 
 unsigned int SEHistogram::NumberOfBoundaries() const
 {
   if (IsValid())
-    return m_Independent.size();
+    return static_cast<unsigned int>(m_Independent.size());
   return 0;
 }
