+++ projects/biogears/src/cdm/utils/FileUtils.cpp
@@ -10,13 +10,13 @@ CONDITIONS OF ANY KIND, either express or implied. See the License for the
 specific language governing permissions and limitations under the License.
 **************************************************************************************/
 
-#include <biogears/cdm/stdafx.h>
+
 #include <biogears/cdm/utils/FileUtils.h>
 #include <dirent.h>
 
 std::string Replace(const std::string& original, const std::string& replace, const std::string& withThis)
 {
-  int idx = 0;
+  size_t idx = 0;
   std::string s = original;
   idx = s.find(replace);
   if (idx != std::string::npos) {
@@ -55,7 +55,7 @@ bool CreateFilePath(const std::string& path)
   std::string dir;
 
   // Create the folders iteratively, backwards
-  for (unsigned int i = folderLevels.size() - 1; i >= 1; i--) {
+  for (size_t i = folderLevels.size() - 1; i >= 1; i--) {
     dir = folderLevels.at(i);
     if (dir == "/" || dir == "\\")
       continue;
@@ -72,7 +72,7 @@ void ListFiles(const std::string& dir, std::vector<std::string>& files, const st
   std::string filename;
   if ((d = opendir(dir.c_str())) != nullptr) {
     while ((ent = readdir(d)) != nullptr) {
-      int nameLength = strlen(ent->d_name);
+      size_t nameLength = strlen(ent->d_name);
 
       if (ent->d_name[0] == '.' && ((nameLength == 1) || (nameLength == 2 && ent->d_name[1] == '.')))
         continue;
@@ -97,7 +97,7 @@ void DeleteDirectory(const std::string& dir, bool bDeleteSubdirectories)
   std::string filename;
   if ((d = opendir(dir.c_str())) != nullptr) {
     while ((ent = readdir(d)) != nullptr) {
-      int nameLength = strlen(ent->d_name);
+      size_t nameLength = strlen(ent->d_name);
 
       if (ent->d_name[0] == '.' && ((nameLength == 1) || (nameLength == 2 && ent->d_name[1] == '.')))
         continue;
